Simplify tests & create mock journal type to check verifability
