+++ crates/light-client-prover/src/circuit.rs
@@ -105,7 +105,10 @@ pub fn run_circuit<DaV: DaVerifier, G: ZkvmGuest>(
                                 &batch_proof_method_id.into(),
                             ) {
                                 Ok(output) => output,
-                                Err(_) => continue,
+                                Err(e) => {
+                                    println!("error: {:?}", e);
+                                    continue;
+                                }
                             };
 
                         // Do not add if last l2 height is smaller or equal to previous output
