+++ crates/sovereign-sdk/adapters/mock-zkvm/src/lib.rs
@@ -46,8 +46,17 @@ pub struct MockProof {
     pub program_id: MockCodeCommitment,
     /// Whether the proof is valid.
     pub is_valid: bool,
-    /// The tamper-proof outputs of the proof.
+    /// The tamper-proof outputs of the proof. Serialized Mock Journal
     pub log: Vec<u8>,
+}
+
+/// A mock journal generated by a zkVM.
+#[derive(Debug, Clone, PartialEq, Eq, BorshDeserialize, BorshSerialize, Serialize, Deserialize)]
+pub enum MockJournal {
+    /// Represents the proof of this journal is verifiable
+    Verifiable(Vec<u8>),
+    /// Represents the proof of this journal is non-verifiable
+    Unverifiable(Vec<u8>),
 }
 
 impl MockProof {
@@ -244,7 +253,11 @@ impl sov_rollup_interface::zk::Zkvm for MockZkGuest {
         journal: &[u8],
         _code_commitment: &Self::CodeCommitment,
     ) -> Result<T, Self::Error> {
-        Ok(T::try_from_slice(journal)?)
+        let mock_journal = MockJournal::try_from_slice(&journal).unwrap();
+        match mock_journal {
+            MockJournal::Verifiable(journal) => Ok(T::try_from_slice(&journal)?),
+            MockJournal::Unverifiable(_) => Err(anyhow::anyhow!("Journal is non-verifiable")),
+        }
     }
 }
